requirements.txt
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8 